+++ packages/lisk-p2p/src/blockchain-p2p/blockchain_peer.ts
@@ -14,7 +14,7 @@
  */
 /* tslint:disable:interface-name no-empty-interface */
 
-import { IPeerConfig, Peer } from '../generic-p2p/peer';
+import { IPeerConfig, Peer } from '../generic-p2p';
 
 export interface IBlockchainPeerConfig extends IPeerConfig {
 	readonly height: number;
